fix: fix tp overlap
